+++ addOns/pscanrules/src/main/java/org/zaproxy/zap/extension/pscanrules/CharsetMismatchScanner.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and 
  * limitations under the License. 
  */
-package org.zaproxy.zap.extension.pscanrulesBeta;
+package org.zaproxy.zap.extension.pscanrules;
 
 import java.util.List;
 
@@ -45,7 +45,7 @@ public class CharsetMismatchScanner extends PluginPassiveScanner {
 	/**
 	 * Prefix for internationalized messages used by this rule
 	 */
-	private static final String MESSAGE_PREFIX = "pscanbeta.charsetmismatch.";
+	private static final String MESSAGE_PREFIX = "pscanrules.charsetmismatch.";
 	
 	private static enum MismatchType {
 		NO_MISMATCH_METACONTENTTYPE_MISSING,